requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.1.1
blinker==1.4
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.6
cryptography==36.0.2
cssselect==1.2.0
cssutils==2.9.0
et-xmlfile==1.1.0
Faker==13.15.1
Flask==1.1.4
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpath==0.82
kaitaistruct==0.9
ldap3==2.9.1
loguru==0.4.0
lxml==4.9.3
MarkupSafe==2.1.3
mitmproxy==8.0.0
mitmproxy_wireguard==0.1.23
msgpack==1.0.7
openpyxl==3.0.10
packaging==23.2
paramiko==2.11.0
passlib==1.7.4
pluggy==1.3.0
premailer==3.10.0
protobuf==3.19.6
psycopg2==2.9.5
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.5.1
pycparser==2.21
pycryptodome==3.19.0
pydivert==2.1.0
PyMySQL==0.9.3
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pytest==7.1.3
pytest-rerunfailures==13.0
python-dateutil==2.8.2
PyYAML==6.0
redis==5.0.1
requests==2.26.0
requests-toolbelt==0.10.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
six==1.16.0
sortedcontainers==2.4.0
tomli==2.0.1
tornado==6.4
urllib3==1.26.18
urwid==2.1.2
Werkzeug==1.0.1
win32-setctime==1.1.0
wsproto==1.1.0
yagmail==0.15.280
zstandard==0.17.0
